refactor(app): extract mixins and CLI entry point
- PanelMixin: panel split/close/navigate/resize via MRO dispatch
- ToolBridgeMixin: thread-safe widget mounting for tool execution
- cli.py: argparse + headless/TUI launch (entry point moved from app.py)
- pyproject.toml: entry point updated to wingman.cli:main
- License field updated to PEP 639 format

app.py: 826 -> 599 lines
